Refactor (first pass - tests failing)
