+++ src/utils/helpers.js
@@ -1,9 +1,5 @@
 /* eslint-disable no-param-reassign */
 const qs = require('qs');
-const CRC32 = require('crc-32');
-const store = require('./store');
-
-const purchaseEventStorageKey = '_constructorio_purchase_order_ids';
 
 const utils = {
   ourEncodeURIComponent: (str) => {
@@ -77,41 +73,6 @@ const utils = {
     }
 
     return {};
-  },
-
-  hasOrderIdRecord(orderId) {
-    const purchaseEventStorage = JSON.parse(store.session.get(purchaseEventStorageKey));
-    const orderIdHash = CRC32.str(orderId.toString());
-
-    if (purchaseEventStorage && purchaseEventStorage[orderIdHash]) {
-      return true;
-    }
-
-    return null;
-  },
-
-  addOrderIdRecord(orderId) {
-    let purchaseEventStorage = JSON.parse(store.session.get(purchaseEventStorageKey));
-    const orderIdHash = CRC32.str(orderId.toString());
-
-    if (purchaseEventStorage) {
-
-      // If the order already exists, do nothing
-      if (purchaseEventStorage[orderIdHash]) {
-        return;
-      }
-
-      purchaseEventStorage[orderIdHash] = true;
-    } else {
-
-      // Create a new object map for the order ids
-      purchaseEventStorage = {
-        [orderIdHash]: true,
-      };
-    }
-
-    // Push the order id map into session storage
-    store.session.set(purchaseEventStorageKey, JSON.stringify(purchaseEventStorage));
   },
 };
